--- modules/base-zio/src/main/scala/io/janstenpickle/trace4cats/base/context/zio/instances.scala
@@ -1,3 +1,0 @@
-package io.janstenpickle.trace4cats.base.context.zio
-
-object instances extends ZIOContextInstances
