+++ modules/base-zio/src/main/scala/trace4cats/context/zio/instances.scala
@@ -1,0 +1,3 @@
+package trace4cats.context.zio
+
+object instances extends ZIOContextInstances
