+++ modules/inject-zio/src/main/scala/trace4cats/zio/ZIOTraceInstances.scala
@@ -1,0 +1,16 @@
+package trace4cats.zio
+
+import trace4cats.Span
+import zio.blocking.Blocking
+import zio.clock.Clock
+import zio.{Has, RIO}
+
+trait ZIOTraceInstances {
+  implicit val spannedRIOTrace: SpannedRIOTracer = new SpannedRIOTracer
+}
+
+trait ZIOHasTraceInstances {
+  implicit def spannedEnvRIOTrace[
+    R <: Clock with Blocking with Has[Span[RIO[Clock with Blocking, *]]]
+  ]: SpannedEnvRIOTracer[R] = new SpannedEnvRIOTracer[R]
+}
